NewWeb/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
defusedxml==0.7.1
Django==5.2.3
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cleanup==9.0.0
django-embed-video==1.4.10
django-filter==25.1
django-js-asset==3.1.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.1
django-summernote==0.8.20.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
joblib==1.5.1
laces==0.1.2
numpy==2.3.1
openpyxl==3.1.5
pillow==11.2.1
pillow_heif==0.21.0
psycopg2-binary==2.9.10
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2024.2
urllib3==2.5.0
webencodings==0.5.1
Willow==1.10.0
